TS-41945 Splitted up TeamscalePluginTest
