+++ teamscale-gradle-plugin/src/test/kotlin/com/teamscale/TeamscalePluginTestBase.kt
@@ -1,0 +1,74 @@
+package com.teamscale
+
+import com.teamscale.plugin.fixtures.TeamscaleConstants
+import com.teamscale.plugin.fixtures.TestRootProject
+import com.teamscale.test.commons.TeamscaleMockServer
+import org.gradle.testkit.runner.BuildResult
+import org.gradle.testkit.runner.GradleRunner
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.io.TempDir
+import java.io.File
+import java.lang.management.ManagementFactory
+
+
+/**
+ * Integration tests for the Teamscale Gradle plugin.
+ */
+abstract class TeamscalePluginTestBase {
+
+	protected lateinit var teamscaleMockServer: TeamscaleMockServer
+
+	@BeforeEach
+	fun startFakeTeamscaleServer() {
+		teamscaleMockServer = TeamscaleMockServer(TeamscaleConstants.PORT)
+			.withAuthentication(TeamscaleConstants.USER, TeamscaleConstants.ACCESS_TOKEN)
+			.acceptingReportUploads()
+			.withImpactedTests("com/example/project/JUnit4Test/systemTest")
+	}
+
+	@AfterEach
+	fun serverShutdown() {
+		teamscaleMockServer.shutdown()
+	}
+
+	/** The Gradle project in which the simulated checkout and test execution will happen. */
+	lateinit var rootProject: TestRootProject
+
+	@BeforeEach
+	fun setup(@TempDir tempDir: File) {
+		rootProject = TestRootProject(tempDir)
+	}
+
+	protected fun run(vararg arguments: String): BuildResult {
+		return buildRunner(*arguments).build()
+	}
+
+	protected fun runExpectingError(vararg arguments: String): BuildResult {
+		return buildRunner(*arguments).buildAndFail()
+	}
+
+	private fun buildRunner(vararg arguments: String): GradleRunner {
+		val runnerArgs = arguments.toMutableList()
+		val runner = GradleRunner.create()
+		runnerArgs.add("--stacktrace")
+
+		if (ManagementFactory.getRuntimeMXBean().inputArguments.toString()
+				.contains("-agentlib:jdwp") && arguments.contains("unitTest")
+		) {
+			runner.withDebug(true)
+			runner.forwardOutput()
+			runnerArgs.add("--refresh-dependencies")
+			runnerArgs.add("--debug")
+			runnerArgs.add("--debug-jvm")
+		}
+
+		runner
+			.withProjectDir(rootProject.projectDir)
+			.withPluginClasspath()
+			.withArguments(runnerArgs)
+			.withGradleVersion("8.4")
+
+		return runner
+	}
+}
